requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.7
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
Pillow>=10.0.0
tqdm>=4.65.0
tensorboard>=2.14.0
albumentations>=1.3.1
einops>=0.7.0
diffusers>=0.21.0
accelerate>=0.24.0
scipy>=1.11.0
grad-cam>=1.4.8